requirements.txt
boto3==1.35.90
fastapi==0.115.8
Flask==3.1.0
Pillow==11.1.0
protobuf==5.29.3
pymongo==4.10.1
python-dotenv==1.0.1
python-telegram-bot==21.10
Requests==2.32.3
typing_extensions==4.12.2
uvicorn==0.34.0
pytz==2024.2
schedule==1.2.2
gunicorn==23.0.0
pyTelegramBotAPI==4.14.0
google-genai==1.1.0
pydantic==2.10.6
pydantic_core==2.27.2
openai==1.62.0
langchain==0.2.0
langchain-text-splitters==0.2.0